# Use `VECTOR_PTR_RO()` in equality helpers