--- a/src/equal.c
+++ b/src/equal.c
@@ -423,7 +423,7 @@ static inline int dbl_equal_na_scalar(const double* x);
 static inline int cpl_equal_na_scalar(const Rcomplex* x);
 static inline int raw_equal_na_scalar(const Rbyte* x);
 static inline int chr_equal_na_scalar(const SEXP* x);
-static inline int list_equal_na_scalar(SEXP x, R_len_t i);
+static inline int list_equal_na_scalar(const SEXP* x);
 static inline int df_equal_na_scalar(SEXP x, R_len_t i);
 
 // If `x` is a data frame, it must have been recursively proxied
@@ -440,7 +440,7 @@ int equal_na(SEXP x, R_len_t i) {
   }
 
   switch (vec_proxy_typeof(x)) {
-  case vctrs_type_list: return list_equal_na_scalar(x, i);
+  case vctrs_type_list: return list_equal_na_scalar(VECTOR_PTR_RO(x) + i);
   case vctrs_type_dataframe: return df_equal_na_scalar(x, i);
   default: break;
   }
@@ -457,20 +457,6 @@ int equal_na(SEXP x, R_len_t i) {
                                                           \
     for (R_len_t i = 0; i < size; ++i, ++p_x) {           \
       p_out[i] = SCALAR_EQUAL_NA(p_x);                    \
-    }                                                     \
-                                                          \
-    UNPROTECT(2);                                         \
-    return out;                                           \
-  }                                                       \
-  while (0)
-
-#define EQUAL_NA_BARRIER(SCALAR_EQUAL_NA)                 \
-  do {                                                    \
-    SEXP out = PROTECT(Rf_allocVector(LGLSXP, size));     \
-    int* p_out = LOGICAL(out);                            \
-                                                          \
-    for (R_len_t i = 0; i < size; ++i) {                  \
-      p_out[i] = SCALAR_EQUAL_NA(x, i);                   \
     }                                                     \
                                                           \
     UNPROTECT(2);                                         \
@@ -495,7 +481,7 @@ SEXP vctrs_equal_na(SEXP x) {
   case vctrs_type_complex:   EQUAL_NA(Rcomplex, COMPLEX_RO, cpl_equal_na_scalar);
   case vctrs_type_raw:       EQUAL_NA(Rbyte, RAW_RO, raw_equal_na_scalar);
   case vctrs_type_character: EQUAL_NA(SEXP, STRING_PTR_RO, chr_equal_na_scalar);
-  case vctrs_type_list:      EQUAL_NA_BARRIER(list_equal_na_scalar);
+  case vctrs_type_list:      EQUAL_NA(SEXP, VECTOR_PTR_RO, list_equal_na_scalar);
   case vctrs_type_dataframe: {
     SEXP out = df_equal_na(x, size);
     UNPROTECT(1);
@@ -507,7 +493,6 @@ SEXP vctrs_equal_na(SEXP x) {
 }
 
 #undef EQUAL_NA
-#undef EQUAL_NA_BARRIER
 
 static inline int lgl_equal_na_scalar(const int* x) {
   return *x == NA_LOGICAL;
@@ -537,8 +522,8 @@ static inline int chr_equal_na_scalar(const SEXP* x) {
   return *x == NA_STRING;
 }
 
-static inline int list_equal_na_scalar(SEXP x, R_len_t i) {
-  return Rf_isNull(VECTOR_ELT(x, i));
+static inline int list_equal_na_scalar(const SEXP* x) {
+  return *x == R_NilValue;
 }
 
 static inline int df_equal_na_scalar(SEXP x, R_len_t i) {
@@ -622,26 +607,6 @@ do {                                                      \
 }                                                         \
 while (0)
 
-#define EQUAL_NA_COL_BARRIER(SCALAR_EQUAL_NA)  \
-do {                                           \
-  for (R_len_t i = 0; i < p_info->size; ++i) { \
-    if (p_info->p_row_known[i]) {              \
-      continue;                                \
-    }                                          \
-                                               \
-    if (!SCALAR_EQUAL_NA(x, i)) {              \
-      p_out[i] = 0;                            \
-      p_info->p_row_known[i] = true;           \
-      --p_info->remaining;                     \
-                                               \
-      if (p_info->remaining == 0) {            \
-        break;                                 \
-      }                                        \
-    }                                          \
-  }                                            \
-}                                              \
-while (0)
-
 static void vec_equal_na_col(int* p_out,
                              struct df_short_circuit_info* p_info,
                              SEXP x) {
@@ -652,7 +617,7 @@ static void vec_equal_na_col(int* p_out,
   case vctrs_type_complex:   EQUAL_NA_COL(Rcomplex, COMPLEX_RO, cpl_equal_na_scalar); break;
   case vctrs_type_raw:       EQUAL_NA_COL(Rbyte, RAW_RO, raw_equal_na_scalar); break;
   case vctrs_type_character: EQUAL_NA_COL(SEXP, STRING_PTR_RO, chr_equal_na_scalar); break;
-  case vctrs_type_list:      EQUAL_NA_COL_BARRIER(list_equal_na_scalar); break;
+  case vctrs_type_list:      EQUAL_NA_COL(SEXP, VECTOR_PTR_RO, list_equal_na_scalar); break;
   case vctrs_type_dataframe: df_equal_na_impl(p_out, p_info, x); break;
   case vctrs_type_scalar:    Rf_errorcall(R_NilValue, "Can't compare scalars with `vec_equal_na()`");
   default:                   Rf_error("Unimplemented type in `vec_equal_na()`");
@@ -660,4 +625,3 @@ static void vec_equal_na_col(int* p_out,
 }
 
 #undef EQUAL_NA_COL
-#undef EQUAL_NA_COL_BARRIER