--- a/src/equal.h
+++ b/src/equal.h
@@ -155,8 +155,8 @@ static inline int chr_equal_scalar(const SEXP* x, const SEXP* y, bool na_equal) 
   }
 }
 
-static inline bool list_equal_missing_scalar(SEXP x, R_len_t i) {
-  return VECTOR_ELT(x, i) == R_NilValue;
+static inline bool list_equal_missing_scalar(SEXP x) {
+  return x == R_NilValue;
 }
 static inline int list_equal_scalar_na_equal(SEXP x, R_len_t i, SEXP y, R_len_t j) {
   const SEXP xi = VECTOR_ELT(x, i);