Closes #1247 

A very satisfying PR! Many of the macros specific to lists have been removed since we can now use `VECTOR_PTR_RO()`.

Updated equality helpers used in:
- `vec_equal()`
- `vec_equal_na()`
- Dictionary functions
- `vec_identify_runs()`

Generally improves performance with lists a little bit

```r
library(vctrs)

x <- rep(list("a"), 1e5)
```

```r
# before
bench::mark(vec_equal(x, x), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal(x, x)    872µs    997µs      971.     616KB     6.85

# after
bench::mark(vec_equal(x, x), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal(x, x)    550µs    626µs     1547.     616KB     10.9
```

```r
# before
bench::mark(vec_equal_na(x), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal_na(x)    352µs    447µs     2288.     395KB     13.8

# after
bench::mark(vec_equal_na(x), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_equal_na(x)   50.3µs   73.1µs    10872.     395KB     65.6
```

```r
# before
bench::mark(vec_identify_runs(x), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)   1.52ms   1.64ms      582.     397KB     4.11

# after
bench::mark(vec_identify_runs(x), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_identify_runs(x)   1.05ms   1.13ms      861.     397KB     6.07
```

```r
# before
bench::mark(vec_match(x, x), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_match(x, x)   5.59ms   6.11ms      159.    1.66MB     4.93

# after
bench::mark(vec_match(x, x), iterations = 1000)
#> # A tibble: 1 x 6
#>   expression           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_match(x, x)   4.82ms   5.24ms      189.    1.66MB     5.85
```